Excludes builds failing due to code/compiler issues
CCE 13.0.0.1 requires the use of `-Wno-unsupported-target-opt` compiler
flags which breaks PETSc's build system tests identifying the MPI
fortran compiler. The need to use this compiler flag appears to be fixed
in CCE 13.0.0.5-beta.

Slepc exclusions for CCE builds are made more generic since the fortran
lib configuration error in Slepc occurs under a wide range of CCE
toolchain versions. The Slepc%cce exception is made specifc to slepc
versions so that future updates to the slepc package which may fix this
error are not excluded but the exception is made less specific to the
CCE toolchain used for a given build.

CCE 13.0.0.5-beta hits an ICE regarding scalar evolution pre-processing
while building ifpack2 components of trilinos. While we could attempt to
build trilinos without this option, we will wait to see if the problem
is fixed in the GA release of CCE 13.0.0.5 before altering the package
variants and specs. Until then, builds of trilinos and its dependents
(omega-h, fortrilinos, datatransferkit) are excluded for CCE
13.0.0.5-beta.