Adds mechanism for settng custom moduleroot directory name to init script
By initializing the spack runtime environment with a variable
`FS_CUSTOM_MODULE_DIR` set, we can now setup out-of-band modulefile
trees that have both the static site modules installed that preserve the
functionality of the static modulefiles' relative MODULEPATH
modifications that integrate with spack-generate modules.

This feature was first added to support an out-of-band preview
modulefile tree for the 2021.11 Andes upgrade.

This commit also reverts the modulefile roots changed in the Andes
spack.yaml configuration file that was made to prevent software built
for the preview upgrade nodes to be exposed to general users on the
yet-to-be-upgraded production Andes nodes.