Enforces use of specific PMI library to overcome cray-pmi v6.0.15 bug
Cray-PMI v6.0.15 was, initally, the default on Crusher and attempts to
fix a known issue with MPMD job launch but introduces new issues.

As of 2021-11-29, v6.0.14 is now the default. This commit is a
work-around to use v6.0.14 while the newer version was the default. It
reverts to using the Cray compiler driver wrappers for MPI codes even if
"vanilla" mpich wrappers are available and adds the previous v6.0.14 PMI
libs to the `LD_LIBRARY_PATH` at build time.

It is intended to revert the addition of specific PMI libs to the
LD_LIBRARY_PATH and revert back to using the vanilla MPI wrappers once
issues with Cray-PMI have been fixed in the next CPE release (ie, revert
this commit and the changes made to the cray-mpich package in the OLCF
fork of spack committed here). Alternatively, if it will continue to be
necessary to add the PMI libs to the build-time `LD_LIBRARY_PATH` when
using either the vanilla or cray MPI compiler drivers, that modification
should be incorporated into the cray-mpich spack package in a future
commit.