Targets gfx90a where possible
Where possible, we target the gfx90a GPU architecture on Frontier. The
Kokkos package is a notable exclusion from this change and must be
updated to v4.5.0 (diverging from the prescribed E4S release). To do
this, major changes must be made to the Kokkos package that are not yet
in upstream spack.

Also, MAGMA fails to build for gfx90a prior to ROCm 4.5.2.