Replaces site modulefile deployment method from cp to rsync
Deploying the site modulefiles now uses rsync to copy out only tracked
site modulefiles that have changed based on file checksum rather than
size or timestamps. Due to the way file timestamps are not preserved in
git, selecting modules to update based on checksum results in fewer
prompts to copy out files that really haven't changed.

An rsync itemized summary of the files to be deployed is shown prior to
prompting the user if they want to deploy the changes.

This change still can be improved by:
- Adding a way to bypass the user prompt for non-interactive uses of
  this script (eg, in CI jobs)
- Checking that rsync is available and falling back to `cp`-based
  deployments when it is not.
- Possibly checking the sent/recieved/created rsync operation mode for
  directory creation operations instead of just "send" (">") operations.
- Adding a method for a user to selectively copy out modulefiles with
  `cp` if a bulk transfer operation is undesireable.
- Relocating the mktemp file to the `$CWD`; though this might not be
  necessary.
- Adding a trap to clean up temp files or replacing the tmp file with a
  bash/zsh portable input stream.

This operation is sensitive to bash/zsh portability issues since it uses
`read` which is a shell builtin function and not a POSIX command.