Partial update of Borg environment
Borg is offline at the momement. But when it is restored, we expect to
have the December PE available. This commit updates the borg environment
configuration to what will be expected when access is restored. It will
need to be checked for path and module correctness before
re-concretizing and deploying software.

It provisionally adds PrgEnv-amd builds, removes obsolete ROCm external
references, and updates the CCE toolchains used for builds. Furthermore,
it reconfigures target architectures to gfx90a and zen3 wherever
possible.