Adds patch for NVHPC support in SMPI compiler wrapper
This patch (or similar) currently needs to be applied post-install
manually for `spectrum-mpi@10.4.0.3-20210112%nvhpc` and earlier builds.
The next version update of SMPI added to the olcf.spectrum-mpi spack
package should be made to incorporate this patch automatically. But
retroactively adding it to the spack package SpectrumMpi class will
cause undue disruption to existing software installations.

It is added to this repo now to keep a tracked record of it.

This patch modifies the SMPI opal_wrapper script (backend to the user
interface `mpicc`, `mpic++`, `mpifort`, etc. compiler wrappers) to be
aware of nvc and nvfortran compilers. The updated opal_wrapper script
will, when using nvhpc, apply the MPI fortran modules and libaries built
to be compatible with PGI (and, by extension, nvhpc) (ie
`fortsupport_PGI`) to compiler invocations.