Initial commit of SMPI fortran support module rebuild script.
This script, at this commit, is an untested script that captures the
manual operations needed to rebuild the spectrum-mpi fortsupport modules
for spectrum-mpi@10.4.0.3-20210112 and later. It is intended to be
useable for rebuilding the fortsupport modules under GCC and NVHPC
toolchains. However, it is - again - not tested yet.

To date, rebuilding the fortsupport modules provided with an SMPI
release has been a manually post-install operation. Ultimately, the
functionality of this script should be incorporated into the
olcf.spectrum-mpi spack package and done automatically as-needed when a
new spectrum-mpi build is installed.

However, due to breaking redesigns and organizational changes to SMPI's
fortsupport module source code each SMPI release, it has been difficult
to automate this procedure prior to installing SMPI and it has been
relagated to a manual post-install operation. Furthermore, these
operations are only needed if the stock fortsupport modules and
libraries provided with SMPI for various toolchains are not sufficient.
Since v10.4.0.3, rebuilding the fortsupport modules is very rarely
necessary.

Until it has been tested, this script should be considered just a guide
tracked in this git repo for posterity meant to illustrate the
operations that have been performed manually when installing SMPI for
gcc<8 and certain pgi/nvhpc toolchains.