Adds placeholder entry for staging on GPFS
Due to cgroup memory limits, some large packages (eg, `qt`) need to be
staged from GPFS to avoid exceeding the limit and having builds be
violently terminated. Hitting the memory limit is more likely when
failed builds accumulate when staging in `/tmp`. However, it is
preferably to stage from `/tmp` when possible. Some packages, such as
`m4` have build systems that are confused when building from GPFS and
must be staged in `/tmp` or an NFS mount.

This commit adds a placeholder entry in the environment's build_stage
configuration paths that should be uncommented and edited as needed to
successfully stage and build problem packages but returned to using
`/tmp` by default when staging from GPFS is not necessary.