Initial commit of AFW spack environment
This commit adds an initial work-in-progess spack environment for the
AFW machines. Notably, the olcf-spack-environment init script is updated
to handle cray systems using TCL environment modules. However, since the
AFW machines are using TCL environment modules, spack build
environments on the AFW systems are inconsistent and not ready for
production. The AFW spack.yaml file added in this commit is merely a
starting point for establishing a spack configuration capable of
building software targetting the Cray platform for the AFW systems and
is not production ready as of this commit.