Adds utility to refresh only missing modulefiles
This commit adds a script which finds all packages explicitly installed
from a spack environment since a given date and refreshes modulefiles
only for builds which do not have a modulefile already publicly
available. That is, this script generates modules for packages that did
not get one due to a build error that occured during a full spack
environment deployment without forcibly refreshing modulefiles
that already exist as would be done if simply executing

```
spack module lmod refresh -y
```

and updating modules for all packages.

The purpose of this script is to allow missing modulefiles to be
deployed for new builds while preserving the currently available modules
and minimizing disruptive impact to users actively using
facility-installed software.

This is a first draft of this script. Ideally, this script's
functionality would be added to the OLCF spack command extension to
improve speed. This script also needs improvements to narrow down the
list of builds targetted for refreshing; currently it looks only for
packages installed since a hardcoded date. This date window and other
selection criteria should be adjustable by the user dynamically when the
script is used.