Adds GCC 11.2.0 and software to Ascent
Installed to satisfy [OLCFSW-180](https://jira.ccs.ornl.gov/projects/OLCFSW/queues/custom/131).

GCC installed manually using `/sw/ascent/gcc/.install/build-11.2.0-0.sh`.

Spectrum-MPI installed first by itself and then updated manually to
rebuild MPI fortran support using the script tracked in this repo:

```
./utils/smpi_rebuild_gfortran.sh gcc/11.2.0 spectrum-mpi/10.4.0.3-20210112
```

Subsequently the rest of the package specs to be built with GCC11.2.0
were added to the ascent environment file; the environment
reconcretized; and specs installed. This is to say that this commit, by
itself, does not represent the order of operations used to actually
deploy GCC 11.2.0 and the packages built with it.

The modulefile for GCC 11.2.0 (tracked in this repo) is hidden via LMOD
RC configs until all the packages to be built are installed.