Initial commit of E4S reference specs
The files added in this commit are reference lists of E4S specs used on
OLCF systems. The authoritative specs are, as always, what are in the
system's `spack.yaml` file. These files are used to construct spec lists
usable in OLCF spack.yaml files from the upstream E4S release manifests
found at "https://github.com/E4S-Project/e4s".

The specs contained in these files are in the form of definition lists
which are intended to be used in spec matrixes that expand the E4S
package root spec across a variety of compilers and runtime dependencies
(eg, CUDA and ROCm) on OLCF systems.

Prior to E4S v21.11, the specs contained in these files are *condensed*
lists that combine both specs that target the CPU and specs
which target the GPUs. After migrating to use the new Clingo
concretizer, this format was inconvenient since spack matrixes under the
clingo concretizer do not permit mixing partial specs in a definition
list with partial dependency specs on which the root spec does not
depend (a behavior that departs from the original spack concretizer).
Therefore, the files in this commit for E4S v20.10 to v21.08 are in a
loose format which is not directly usable in OLCF spack.yaml files.
They are included only as a reference and a guide for what may have been
installed on an OLCF machine in the past.

From E4S 21.11 and onward, the definition lists in these files separate
CPU-only specs from GPU-capable specs such that spec matrixes in a
spack.yaml file may combine GPU-capable specs with specific CUDA or ROCm
implementations needed for use with specific toolchains and to
facilitate easier package rebuilds for GPU runtime upgrades. In
practice, only the minimally different specs in a new E4S release are
added to a production spack.yaml file to minimize the time to solve the
environments and minimize spurious rebuilds of packages which are
already installed. Therefore, the files in this commit for E4S v21.11+
also include minimal diff spec lists from the previous E4S release.
The v21.11+ files also use a more refined system of commentted-out specs
to indicate:

- Packages should be installed manually or as a single build under
  one toolchain (ie, not expanded in spec matrixes for all compilers).
  Denoted by a single (1) comment hash token.
- Packages which exist in previous E4S releases and have been built
  successfully in the past which do not generally need to be added to a
  machine that is up-to-date. This token is not widely used.
  Denoted by two (2) comment hash tokens.
- Packages that exist in a previous E4S release which have not been
  built successfully due to an unexpected build error or problem with
  the code. Such packages should not be added to a spec.yaml file
  without testing after version upgrades or changes to the system which
  may have resolved the previous build error.
  Denoted by three (3) comment hash tokens.
- Packages which are officially in an E4S release but have been rejected
  by the OLCF for facility deployments. Such packages typically have
  dependencies or capabilities which are incompatible with an OLCF
  target machine (eg, a package needing a scheduler not available on the
  system), or they violate facility software deployment policies and
  SLAs (eg, python extension packages), or they are packages which the
  OLCF staff or vendors install using means other than spack (eg,
  compilers, graphical visualization tools, debuggers/profilers, or
  software needing system-level configuration such as container runtimes).
  Denoted by four (4) comment hash tokens.