WIP: Updates summit env with newly hashed SMPI rebuilds
The SMPI spack package needs updated build/run environment methods. Old
builds of SMPI are satisfactory, and should continue to be used when
concretizing an environment under spack v0.18 s `--reuse` (default)
concretization flag. But if using the spack v0.18 `--fresh/--force`
flags during concretization, SMPI will be necessarily get a new hash and
be re-installed.

A change to a spec in the "$general_compute_packages" definition list on
Summit requires some packages to be rebuilt using the `--fresh` flag
prior to solving the environment for new NVHPC 22.5 software builds.
This commit updates the SMPI spack package (and other OLCF spackage repo
packages) to use the new build/run environment package API methods, and
updates the summit environment yaml file, and lock file to update
package builds for existing (ie, not nvhpc 22.5) compilers that are
affected by the general_compute_packages spec change and would not be
able reuse existing SMPI builds so we can have a stable reset point for
the lock file to revert to when debugging nvhpc 22.5 builds in the
immediate future.