Initial commit of utility command for dumping environment specs
For the purpose of reporting installed packages in a compact
standardized format, a utility script has been added to call `spack
find` in a consistent output format.

This initial commit uses a format that has been used in the past for ECP
reporting but it is a crude format which could stand improvement.