Updates Summit CUDA-dependent specs for breaking change to packages
Cabana and Arborx can optionally depend on CUDA but at least one of
these pacakges did not seem to propogate the cuda target architecture
correctly to its kokkos dependency causing concretization errors. Using

The GPU-enabled spec lists were checked against the list provided by

```
spack dependents kokkos
```

to determine which specs in the environment might need a specific kokkos
dependency spec element added to the root. This commit then forces the
kokkos dependency for CUDA-enabled builds of cabana to target the
correct GPU architecture. It also explicitly adds a previously missing
cuda_arch target to arborx specs in the environment. For now, this
change is applied only to Summit but will need to be propagated to Peak
and Ascent.

The issue of missing `cuda_arch` targets has come up in the past. A new
feature proposal to spack for a default `cuda_arch` target that can be
set globally in an environment for `CudaPackage` class specs that
currently resolve by default to `+cuda cuda_arch=none` when only `+cuda`
is set at a spec DAG node would be useful.