Adds spack extension for refreshing missing modulefiles
This commit replaces `/utils/detect_missing_modules.sh` in this repo
with a spack extension command which is orders of magnitude faster. The
command `spack olcf find-missing-modules` scans explicitly installed
packages in the current environment and checks if a modulefile for it
or an existing name-colliding package build exists among the `default`
LMOD module set. Packages for which a current or name-conflicting
module exist are not affected, but packages for which a module does not
exist are identified and the user is given the option to generate
modules for them.