Updates missing module utility default start date
The script to detect missing modules and generate them previously took a
user-overrideable start date which had a hard-coded default value. This
commit replaces the default value with one that is 4 days in the past to
allow time for builds to finish without reaching so far back in time as
to pickup builds that were not freshly added. The default value can
still be overriden by the user.